@@ -1,94 +1,105 @@
 class TransactionsController < ApplicationController
   def index
     @q = Transaction.ransack(params[:q])
-    @transactions = @q.result(:distinct => true).page(params[:page]).per(10)
+    @transactions = @q.result(:distinct => true).includes(:asset, :matter).page(params[:page]).per(10)
 
-    render("transactions/index.html.erb")
+    render("transaction_templates/index.html.erb")
   end
 
   def show
-    @transaction = Transaction.find(params[:id])
+    @transaction = Transaction.find(params.fetch("id_to_display"))
 
-    render("transactions/show.html.erb")
+    render("transaction_templates/show.html.erb")
   end
 
-  def new
+  def new_form
     @transaction = Transaction.new
 
-    render("transactions/new.html.erb")
+    render("transaction_templates/new_form.html.erb")
   end
 
-  def create
+  def create_row
     @transaction = Transaction.new
 
-    @transaction.asset_id = params[:asset_id]
-    @transaction.shares = params[:shares]
-    @transaction.date = params[:date]
-    @transaction.amount = params[:amount]
-    @transaction.category = params[:category]
-    @transaction.description = params[:description]
-    @transaction.payee = params[:payee]
-    @transaction.distribution = params[:distribution]
+    @transaction.asset_id = params.fetch("asset_id")
+    @transaction.shares = params.fetch("shares")
+    @transaction.date = params.fetch("date")
+    @transaction.amount = params.fetch("amount")
+    @transaction.category = params.fetch("category")
+    @transaction.description = params.fetch("description")
+    @transaction.payee = params.fetch("payee")
+    @transaction.distribution = params.fetch("distribution")
 
-    save_status = @transaction.save
+    if @transaction.valid?
+      @transaction.save
 
-    if save_status == true
-      referer = URI(request.referer).path
-
-      case referer
-      when "/transactions/new", "/create_transaction"
-        redirect_to("/transactions")
-      else
-        redirect_back(:fallback_location => "/", :notice => "Transaction created successfully.")
-      end
+      redirect_back(:fallback_location => "/transactions", :notice => "Transaction created successfully.")
     else
-      render("transactions/new.html.erb")
+      render("transaction_templates/new_form_with_errors.html.erb")
     end
   end
 
-  def edit
-    @transaction = Transaction.find(params[:id])
+  def create_row_from_asset
+    @transaction = Transaction.new
 
-    render("transactions/edit.html.erb")
-  end
+    @transaction.asset_id = params.fetch("asset_id")
+    @transaction.shares = params.fetch("shares")
+    @transaction.date = params.fetch("date")
+    @transaction.amount = params.fetch("amount")
+    @transaction.category = params.fetch("category")
+    @transaction.description = params.fetch("description")
+    @transaction.payee = params.fetch("payee")
+    @transaction.distribution = params.fetch("distribution")
 
-  def update
-    @transaction = Transaction.find(params[:id])
+    if @transaction.valid?
+      @transaction.save
 
-    @transaction.asset_id = params[:asset_id]
-    @transaction.shares = params[:shares]
-    @transaction.date = params[:date]
-    @transaction.amount = params[:amount]
-    @transaction.category = params[:category]
-    @transaction.description = params[:description]
-    @transaction.payee = params[:payee]
-    @transaction.distribution = params[:distribution]
-
-    save_status = @transaction.save
-
-    if save_status == true
-      referer = URI(request.referer).path
-
-      case referer
-      when "/transactions/#{@transaction.id}/edit", "/update_transaction"
-        redirect_to("/transactions/#{@transaction.id}", :notice => "Transaction updated successfully.")
-      else
-        redirect_back(:fallback_location => "/", :notice => "Transaction updated successfully.")
-      end
+      redirect_to("/assets/#{@transaction.asset_id}", notice: "Transaction created successfully.")
     else
-      render("transactions/edit.html.erb")
+      render("transaction_templates/new_form_with_errors.html.erb")
     end
   end
 
-  def destroy
-    @transaction = Transaction.find(params[:id])
+  def edit_form
+    @transaction = Transaction.find(params.fetch("prefill_with_id"))
+
+    render("transaction_templates/edit_form.html.erb")
+  end
+
+  def update_row
+    @transaction = Transaction.find(params.fetch("id_to_modify"))
+
+    @transaction.asset_id = params.fetch("asset_id")
+    @transaction.shares = params.fetch("shares")
+    @transaction.date = params.fetch("date")
+    @transaction.amount = params.fetch("amount")
+    @transaction.category = params.fetch("category")
+    @transaction.description = params.fetch("description")
+    @transaction.payee = params.fetch("payee")
+    @transaction.distribution = params.fetch("distribution")
+
+    if @transaction.valid?
+      @transaction.save
+
+      redirect_to("/transactions/#{@transaction.id}", :notice => "Transaction updated successfully.")
+    else
+      render("transaction_templates/edit_form_with_errors.html.erb")
+    end
+  end
+
+  def destroy_row_from_asset
+    @transaction = Transaction.find(params.fetch("id_to_remove"))
 
     @transaction.destroy
 
-    if URI(request.referer).path == "/transactions/#{@transaction.id}"
-      redirect_to("/", :notice => "Transaction deleted.")
-    else
-      redirect_back(:fallback_location => "/", :notice => "Transaction deleted.")
-    end
+    redirect_to("/assets/#{@transaction.asset_id}", notice: "Transaction deleted successfully.")
+  end
+
+  def destroy_row
+    @transaction = Transaction.find(params.fetch("id_to_remove"))
+
+    @transaction.destroy
+
+    redirect_to("/transactions", :notice => "Transaction deleted successfully.")
   end
 end
