@@ -1,15 +1,15 @@
 class ApplicationController < ActionController::Base
   before_action :authenticate_user!
 
-  protect_from_forgery with: :exception
+  # protect_from_forgery with: :exception
 
   before_action :configure_permitted_parameters, if: :devise_controller?
 
   protected
 
   def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:sign_up, :keys => [:first_name, :last_name, :mail_address, :corporate, :firm_name])
+    devise_parameter_sanitizer.permit(:sign_up, :keys => [:first_name, :last_name, :mail_address, :isadmin, :profile_pic])
 
-    devise_parameter_sanitizer.permit(:account_update, :keys => [:first_name, :last_name, :mail_address, :corporate, :firm_name])
+    devise_parameter_sanitizer.permit(:account_update, :keys => [:first_name, :last_name, :mail_address, :isadmin, :profile_pic])
   end
 end
