@@ -1,19 +1,30 @@
 class Matter < ApplicationRecord
   # Direct associations
 
-  has_many   :beneficiaries,
-             :dependent => :destroy
-
-  has_many   :fiduciaries,
+  has_many   :matter_joins,
+             :class_name => "Assignment",
              :dependent => :destroy
 
   has_many   :assets,
              :dependent => :destroy
 
-  belongs_to :user,
-             :counter_cache => true
-
   # Indirect associations
+
+  has_many   :users,
+             :through => :matter_joins,
+             :source => :user
+
+  has_many   :transactions,
+             :through => :assets,
+             :source => :transactions
+
+  has_many   :asset_values,
+             :through => :assets,
+             :source => :asset_values
+
+  has_many   :organizations,
+             :through => :matter_joins,
+             :source => :organization
 
   # Validations
 
